+++ extract_msg/__main__.py
@@ -70,6 +70,7 @@ def main() -> None:
             'saveHeader': args.saveHeader,
             'skipBodyNotFound': args.skipBodyNotFound,
             'skipEmbedded': args.skipEmbedded,
+            'skipHidden': args.skipHidden,
             'skipNotImplemented': args.skipNotImplemented,
             'useMsgFilename': args.useFilename,
             'wkOptions': args.wkOptions,