+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+**v0.37.1**
+* Added option to save function (including `MSGFile.saveAttachments`) to skip any attachments marked as hidden. This can also be done from the command line with `--skip-hidden`.
+* Fixed an issue that could cause an `MSGFile` to be included in the attachment names instead of just strings.
+* Fixed up several comments.
+* Fixed bug that would cause spaces at the end of the subject or filename to break the module.
+
 **v0.37.0**
 * [[TeamMsgExtractor #303](https://github.com/TeamMsgExtractor/msg-extractor/issues/303)] Renamed internal variable that wasn't changed when the other instances or it were renamed.
 * [[TeamMsgExtractor #302](https://github.com/TeamMsgExtractor/msg-extractor/issues/302)] Fixed properties missing (a fatal MSG error) raising an unclear exception. It now uses `InvalidFileFormatError`.